fix(vocabulary): fix bug update word

+++ app/src/main/java/com/example/project_enlishlearning/navigation/NavGraph.kt
@@ -376,8 +376,7 @@ fun AppNavGraph(
             )
         ) { backStackEntry ->
 
-            val wordId =
-                backStackEntry.arguments?.getInt("wordId") ?: 0L
+            val wordId = backStackEntry.arguments?.getLong("wordId") ?: 0L
 
             EditVocabularyScreen(
                 navController = navController,
